Code refactor with readMoreLink.
