+++ js/src/components/google-account-card/read-more-link.js
@@ -1,0 +1,14 @@
+/**
+ * Internal dependencies
+ */
+import AppDocumentationLink from '.~/components/app-documentation-link';
+
+const readMoreLink = (
+	<AppDocumentationLink
+		context="setup-mc-accounts"
+		linkId="required-google-permissions"
+		href="https://docs.woocommerce.com/document/google-listings-and-ads/#required-google-permissions"
+	/>
+);
+
+export default readMoreLink;
